requirements.txt
Flask
flask-cors
flask-mysqldb
Pillow
torch
torchvision
